requirements.txt
flask==2.0.1
werkzeug==2.0.3
slack_sdk==3.19.5
python-dotenv==0.19.1
requests==2.26.0 